Refactor garden application service into focused types
