+++ backend/SurvivalGarden.Application/IGardenStateStore.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Nodes;
+
+namespace SurvivalGarden.Application;
+
+public interface IGardenStateStore
+{
+    Task<JsonObject?> LoadAsync(CancellationToken cancellationToken = default);
+    Task SaveAsync(JsonObject appState, CancellationToken cancellationToken = default);
+}
